requirements.txt
awscli==1.15.26
backcall==0.1.0
boto3==1.7.26
botocore==1.10.26
click==6.7
colorama==0.3.9
decorator==4.3.0
docutils==0.14
imageio==2.3.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jmespath==0.9.3
numpy==1.14.3
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
progressbar2==3.37.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.3
Pygments==2.2.0
python-dateutil==2.7.3
python-utils==2.3.0
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
yapf==0.22.0
